test: migrate search/ package into integration root (PR 2)
Move all 7 context-loading search test classes from com.otilm.core.search
into the single integration root com.otilm.core.integration.search and rename
them *ITest to match the test-integration surefire profile:

  DiscoveryHistorySearchTest            -> integration/search/DiscoveryHistorySearchITest
  EntityInstanceSearchTest              -> integration/search/EntityInstanceSearchITest
  LocationsSearchTest                   -> integration/search/LocationsSearchITest
  NativeArrayFilterSearchTest           -> integration/search/NativeArrayFilterSearchITest
  SigningProfileSearchTest              -> integration/search/SigningProfileSearchITest
  TimeQualityConfigurationSearchTest    -> integration/search/TimeQualityConfigurationSearchITest
  TspProfileSearchTest                  -> integration/search/TspProfileSearchITest

All 7 extend BaseSpringBootTest (pure integration, 7/7) so this is a straight
move+rename with no class splitting. Remove "search" from the CiTestSplitTest
MIGRATION_ALLOWLIST now that the package holds no context test outside the
integration root; the self-prune guard keeps the allowlist honest.

Part of the integration-test extraction series (#1710), following #1700 (PR 0)
and #1709 (PR 1). Epic OmniTrustILM/ilm#252.
