+++ src/test/java/com/otilm/core/architecture/CiTestSplitTest.java
@@ -126,7 +126,7 @@ class CiTestSplitTest {
      * fragments — a substring match would wrongly exempt e.g. any path merely containing "api".
      */
     private static final Set<String> MIGRATION_ALLOWLIST = Set.of(
-            "service", "util", "signing", "messaging", "security", "search",
+            "service", "util", "signing", "messaging", "security",
             "dao", "api", "attribute", "config", "tasks", "events", "repository", "auth",
             "aop", "model", "provisioning", "evaluator", "cryptography", "cluster");
