+++ src/test/java/com/otilm/core/integration/search/TspProfileSearchITest.java
@@ -1,4 +1,4 @@
-package com.otilm.core.search;
+package com.otilm.core.integration.search;
 
 import com.otilm.api.model.client.attribute.RequestAttributeV3;
 import com.otilm.api.model.client.certificate.SearchFilterRequestDto;
@@ -38,7 +38,7 @@ import static com.otilm.core.util.builders.SearchFilterRequestDtoBuilder.aProper
 import static com.otilm.core.util.builders.SearchFilterRequestDtoBuilder.aPropertyFilter;
 import static com.otilm.core.util.builders.SearchFilterRequestDtoBuilder.aPropertyNotEqualsFilter;
 
-class TspProfileSearchTest extends BaseSpringBootTest {
+class TspProfileSearchITest extends BaseSpringBootTest {
 
     private static final String BASE_URL = "http://localhost";
     private static final String CUSTOM_ATTR_NAME = "tsp-tag";
